This is my first project i did for my internship
The project is based on Ecommerce pain in Health domain.The project is based on solving the issue of emergency medicines being delivered late by a pharmacy or any organization to the required patient. My project is based on solving this issue by sorting out medicines into 5 categories,with each category having a priority order of 1,2,3,4 and 5,with 1 being medicines that have highest priority and have to be delivered as soon as possible to the customer/patient,while the priority 5 meds consist of geeneric medicines for low fevers,etc.